+++ src/Resources/Pages/EditRecord/Concerns/Translatable.php
@@ -50,7 +50,6 @@ trait Translatable
         foreach ($this->otherLocaleData as $locale => $localeData) {
             try {
                 $this->form->fill($this->form->getState());
-                $this->form->validate();
             } catch (ValidationException $exception) {
                 if (! array_key_exists($locale, $record->locales())) {
                     continue;
@@ -76,36 +75,6 @@ trait Translatable
     public function updatingActiveLocale(): void
     {
         $this->oldActiveLocale = $this->activeLocale;
-    }
-
-    public function updatedActiveLocale(): void
-    {
-        if (blank($this->oldActiveLocale)) {
-            return;
-        }
-
-        $this->resetValidation();
-        $translatableAttributes = static::getResource()::getTranslatableAttributes();
-        try {
-            $this->otherLocaleData[$this->oldActiveLocale] = Arr::only(
-                $this->form->getState(),
-                $translatableAttributes
-            );
-
-            $this->form->fill([
-                ...Arr::except(
-                    $this->form->getState(),
-                    $translatableAttributes
-                ),
-                ...$this->otherLocaleData[$this->activeLocale] ?? [],
-            ]);
-
-            unset($this->otherLocaleData[$this->activeLocale]);
-        } catch (ValidationException $e) {
-            $this->activeLocale = $this->oldActiveLocale;
-
-            throw $e;
-        }
     }
 
     public function setActiveLocale(string $locale): void
